refactor(core): promote pkg/lifecycle to internal/core/lifecycle (no behavior change)
