+++ cmd/databricks-claude/serve_session.go
@@ -12,10 +12,10 @@ import (
 	"syscall"
 
 	"github.com/IceRhymers/databricks-agents/internal/core/authcheck"
+	"github.com/IceRhymers/databricks-agents/internal/core/lifecycle"
 	"github.com/IceRhymers/databricks-agents/internal/core/portbind"
 	"github.com/IceRhymers/databricks-agents/internal/core/refcount"
 	"github.com/IceRhymers/databricks-agents/pkg/health"
-	"github.com/IceRhymers/databricks-agents/pkg/lifecycle"
 	"github.com/IceRhymers/databricks-agents/pkg/proxy"
 )
