+++ internal/core/lifecycle/lifecycle.go
