+++ internal/core/lifecycle/lifecycle_test.go
